rename IGUIItem
