+++ src/main/java/appeng/api/implementations/menuobjects/IPortableCell.java
@@ -21,26 +21,14 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package appeng.api.implementations.guiobjects;
+package appeng.api.implementations.menuobjects;
 
-import net.minecraft.core.BlockPos;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-
-import javax.annotation.Nullable;
+import appeng.api.networking.energy.IEnergySource;
+import appeng.api.storage.ITerminalHost;
 
 /**
- * Implemented on Item objects, to return objects used to manage, and interact with the contents.
+ * Obtained via {@link IMenuItem} getMenuHost
  */
-public interface IGuiItem {
-    /**
-     * @param player        The player opening the menu.
-     * @param inventorySlot The slot in the player's inventory the item is in.
-     * @param stack         The item used to open the menu.
-     * @param pos           If the item was used on a block in the world, has that block's position. Null if no
-     *                      block was targeted.
-     * @return Null if no GUI object is available.
-     */
-    @Nullable
-    ItemMenuHost getGuiObject(Player player, int inventorySlot, ItemStack stack, @Nullable BlockPos pos);
+public interface IPortableCell extends ITerminalHost, IEnergySource {
+
 }
