+++ src/main/java/appeng/helpers/WirelessCraftingTerminalMenuHost.java
@@ -1,14 +1,16 @@
 package appeng.helpers;
 
+import org.jetbrains.annotations.Nullable;
+
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+
 import appeng.api.inventories.ISegmentedInventory;
 import appeng.api.inventories.InternalInventory;
 import appeng.parts.reporting.CraftingTerminalPart;
 import appeng.util.inv.AppEngInternalInventory;
 import appeng.util.inv.InternalInventoryHost;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import org.jetbrains.annotations.Nullable;
 
 public class WirelessCraftingTerminalMenuHost extends WirelessTerminalMenuHost
         implements ISegmentedInventory, InternalInventoryHost {
